requirements.txt
scipy
NumPy
matplotlib
sklearn
nose
tensorflow-gpu
Pillow
tflearn
jupyter
pandas
seaborn
keras
h5py
python-daemon
scikit-image
docopt
PyYAML
